requirements.txt
apache-airflow==1.10.4
kubernetes==11.0.0
watchdog==0.10.2